rename masked_load to masked_load_store
