+++ triton_viz/core/nki.py
@@ -11,7 +11,7 @@ except (
 import inspect
 import textwrap
 from ..transformers.nki_extract_slice import transform_code
-from .masked_load import masked_load, masked_store
+from .masked_load_store import masked_load, masked_store
 
 
 class NDArray:
